+++ libs/search-widget/src/core/api.ts
@@ -3,6 +3,7 @@ import {
   Chat,
   Classification,
   FieldFullId,
+  IEvents,
   IResource,
   KBStates,
   LabelSets,
@@ -23,7 +24,7 @@ import { suggestionError } from './stores/suggestions.store';
 import { NucliaPrediction } from '@nuclia/prediction';
 import { searchError, searchOptions } from './stores/search.store';
 import { hasViewerSearchError } from './stores/viewer-search.store';
-import { chatError } from './stores/answers.store';
+import { initTracking } from './tracking';
 
 let nucliaApi: Nuclia | null;
 let nucliaPrediction: NucliaPrediction | null;
@@ -31,7 +32,7 @@ let STATE: KBStates;
 let SEARCH_MODE = [Search.Features.PARAGRAPH, Search.Features.VECTOR];
 const DEFAULT_SEARCH_OPTIONS: Partial<SearchOptions> = {};
 
-export const initNuclia = (options: NucliaOptions, state: KBStates, widgetOptions: WidgetOptions): Nuclia => {
+export const initNuclia = (options: NucliaOptions, state: KBStates, widgetOptions: WidgetOptions) => {
   if (nucliaApi) {
     throw new Error('Cannot exist more than one Nuclia widget at the same time');
   }
@@ -45,6 +46,7 @@ export const initNuclia = (options: NucliaOptions, state: KBStates, widgetOption
     DEFAULT_SEARCH_OPTIONS.with_synonyms = true;
   }
   nucliaApi = new Nuclia(options);
+  initTracking(nucliaApi.options.knowledgeBox || 'kb not defined');
   searchOptions.set({
     inTitleOnly: false,
     highlight: widgetOptions.highlight,
@@ -282,4 +284,11 @@ export function getTextFile(path: string): Observable<string> {
     throw new Error('Nuclia API not initialized');
   }
   return nucliaApi.rest.get<Response>(path, {}, true).pipe(switchMap((res) => from(res.text())));
+}
+
+export function getEvents(): IEvents {
+  if (!nucliaApi) {
+    throw new Error('Nuclia API not initialized');
+  }
+  return nucliaApi.events;
 }